# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AfriPebbles API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders/quote:
    post:
      operationId: quoteOrder
      tags: [orders]
      summary: Server-side authoritative price/availability preview for a cart (no order is created)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteOrderInput"
      responses:
        "200":
          description: Computed totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResult"
        "400":
          description: Invalid cart (unavailable product, bad variant, preorder window closed, invalid coupon, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a guest order. Prices/availability are re-validated and computed server-side; client-submitted amounts are never trusted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "400":
          description: Invalid cart or missing consent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/track:
    post:
      operationId: trackOrder
      tags: [orders]
      summary: Look up an order by reference + email (never by id alone)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackOrderInput"
      responses:
        "200":
          description: Limited public order view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicOrderTracking"
        "404":
          description: No matching order for that reference + email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/download:
    get:
      operationId: getOrderItemDownloadUrl
      tags: [orders]
      summary: Mint a short-lived signed download URL for a paid digital order item. Requires the matching email; never returns a permanent public URL.
      parameters:
        - name: orderReference
          in: query
          required: true
          schema:
            type: string
        - name: productId
          in: query
          required: true
          schema:
            type: number
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A signed, time-limited download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadUrlResponse"
        "403":
          description: Order isn't paid yet, or this product isn't a digital item on the order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No matching order for that reference + email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Admin: orders
  # ---------------------------------------------------------------------------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    PaymentMethod:
      type: string
      enum: [paypal, mobile_money, bank_transfer]
    PaymentStatus:
      type: string
      enum: [pending, paid, failed, refunded, cancelled]
    OrderStatus:
      type: string
      enum: [pending_payment, paid, processing, shipped, delivered, cancelled]
...
    OrderItemVariant:
      type: object
      properties:
        label:
          type: string
        option:
          type: string
      required:
        - label
        - option
    CartItemInput:
      type: object
      properties:
        productId:
          type: number
        quantity:
          type: number
        variant:
          oneOf:
            - $ref: "#/components/schemas/OrderItemVariant"
            - type: "null"
      required:
        - productId
        - quantity
    OrderItem:
      type: object
      properties:
        id:
          type: number
        orderId:
          type: number
        productId:
          type: ["number", "null"]
        productName:
          type: string
        productType:
          type: string
          enum: [digital, physical]
        variant:
          oneOf:
            - $ref: "#/components/schemas/OrderItemVariant"
            - type: "null"
        quantity:
          type: number
        unitPrice:
          type: number
        lineTotal:
          type: number
        shippingAmount:
          type: number
        isDigital:
          type: boolean
        isPreorder:
          type: boolean
        preorderFulfilmentText:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - productId
        - productName
        - productType
        - variant
        - quantity
        - unitPrice
        - lineTotal
        - shippingAmount
        - isDigital
        - isPreorder
        - preorderFulfilmentText
        - createdAt
    Order:
      type: object
      properties:
        id:
          type: number
        orderReference:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        country:
          type: string
        deliveryAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        currency:
          type: string
        subtotal:
          type: number
        shippingTotal:
          type: number
        discountTotal:
          type: number
        grandTotal:
          type: number
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        orderStatus:
          $ref: "#/components/schemas/OrderStatus"
        paymentNote:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
          format: date-time
        trackingNumber:
          type: ["string", "null"]
        internalNotes:
          type: ["string", "null"]
        couponCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orderReference
        - customerName
        - customerEmail
        - customerPhone
        - country
        - deliveryAddress
        - notes
        - currency
        - subtotal
        - shippingTotal
        - discountTotal
        - grandTotal
        - paymentMethod
        - paymentStatus
        - orderStatus
        - paymentNote
        - paidAt
        - trackingNumber
        - internalNotes
        - couponCode
        - createdAt
        - updatedAt
    OrderWithItems:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
          required:
            - items
...
    PublicOrderTrackingItem:
      type: object
      properties:
        productId:
          type: ["number", "null"]
        productName:
          type: string
        variant:
          oneOf:
            - $ref: "#/components/schemas/OrderItemVariant"
            - type: "null"
        quantity:
          type: number
        lineTotal:
          type: number
        isDigital:
          type: boolean
        isPreorder:
          type: boolean
      required:
        - productId
        - productName
        - variant
        - quantity
        - lineTotal
        - isDigital
        - isPreorder
    PublicOrderTracking:
      type: object
      properties:
        orderReference:
          type: string
        createdAt:
          type: string
          format: date-time
        currency:
          type: string
        grandTotal:
          type: number
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        orderStatus:
          $ref: "#/components/schemas/OrderStatus"
        trackingNumber:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PublicOrderTrackingItem"
      required:
        - orderReference
        - createdAt
        - currency
        - grandTotal
        - paymentStatus
        - orderStatus
        - trackingNumber
        - items
    TrackOrderInput:
      type: object
      properties:
        orderReference:
          type: string
        email:
          type: string
      required:
        - orderReference
        - email
    CreateOrderInput:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        country:
          type: string
        deliveryAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        couponCode:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItemInput"
        consent:
          type: boolean
      required:
        - customerName
        - customerEmail
        - customerPhone
        - country
        - paymentMethod
        - items
        - consent
    QuoteOrderInput:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItemInput"
        couponCode:
          type: ["string", "null"]
      required:
        - items
    QuoteResultItem:
      type: object
      properties:
        productId:
          type: number
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        lineTotal:
          type: number
        shippingAmount:
          type: number
        isDigital:
          type: boolean
        isPreorder:
          type: boolean
      required:
        - productId
        - productName
        - quantity
        - unitPrice
        - lineTotal
        - shippingAmount
        - isDigital
        - isPreorder
    QuoteResult:
      type: object
      properties:
        currency:
          type: string
        subtotal:
          type: number
        shippingTotal:
          type: number
        discountTotal:
          type: number
        grandTotal:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteResultItem"
      required:
        - currency
        - subtotal
        - shippingTotal
        - discountTotal
        - grandTotal
        - items
    DownloadUrlResponse:
      type: object
      properties:
        downloadUrl:
          type: string
      required:
        - downloadUrl
...