# lib/api-zod/src/generated/types/publicOrderTracking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AfriPebbles API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';
import type { PaymentStatus } from './paymentStatus';
import type { PublicOrderTrackingItem } from './publicOrderTrackingItem';

export interface PublicOrderTracking {
  orderReference: string;
  createdAt: Date;
  currency: string;
  grandTotal: number;
  paymentStatus: PaymentStatus;
  orderStatus: OrderStatus;
  /** @nullable */
  trackingNumber: string | null;
  items: PublicOrderTrackingItem[];
}